Hi, basically I was just messing around and did some benchmarking, found some interesting things, so I made some changes. Mostly to do with removing some of the branches and created a SIMD version of finding CRLF that doesn't need to walk the whole payload twice(worst case)

Below are the benchmark results

### OSX (el capitan, mbp 13" early 2015, 3.1 GHz Intel Core i7)

#### Clang
| time(master) | req/s(master) | time(new) | req/s(new) | percent |
| --- | --- | --- | --- | --- | 
|5.761632|867809.687500|5.104409|979545.312500|11.4069|
|5.769115|866684.062500|5.104001|979623.625000|11.5289|
|5.803461|861554.875000|5.086198|983052.625000|12.3592|
|5.779534|865121.687500|5.082990|983672.937500|12.0519|
|5.762670|867653.375000|5.092542|981827.875000|11.6288|

#### GCC 4.9
| time(master) | req/s(master) | time(new) | req/s(new) | percent |
| --- | --- | --- | --- | --- |
|4.297466|1163476.375000|3.787979|1319965.000000|11.8555|
|4.159397|1202097.250000|3.774015|1324849.000000|9.26533|
|4.230322|1181943.125000|3.788900|1319644.250000|10.4347|
|4.214539|1186369.375000|3.772407|1325413.750000|10.4906|
|4.162938|1201074.750000|3.763118|1328685.375000|9.6042

### ubuntu 14.04 ( Intel Core i7-6700 @ 3.40GHz )

Adding the benchmark in a sec