@shekhei 2016-10-19
I am not sure how to ignore whitespace when you view the diff, the changes are actually very minimal