@indutny 2016-10-19
@shekhei thank you for doing this! I've experimented with vector intrinsics some time before, but for whatever reason I had at the moment decided to not proceed further.

Is the most of the speedup achieved by using vector intrinsics? How much do other changes contribute?