@shekhei 2016-10-20
It seems like they each contribute to approximately half of the speed up. The fallthroughs about 10%, the h_general jmp was about another 20%, the vector speedups is slightly more interesting, it really depends on the size of the header values :)

I was really surprised about the gcc and clang difference though, if i use gcc 6 to compile it, my god its almost twice as fast as the clang one on osx.