@shekhei 2016-10-20
@craig65535 you are right, let me modify them to lower case, and using snake for function names.