@zbjornson 2017-09-16
Eager to see this landed. I get about a 21% speedup on an i7-7700HQ on Windows.

Just tested a wider version (256 bits, requiring AVX2) and get another 8% on top of that. This SSE2-only version needs no dispatching though, which makes it easy. Happy to help with an AVX2 implementation (using either preprocessor directives or dispatching) if there's interest, though.