@shekhei 2017-09-19
Ok, it seems like my whole branch has gone cranky after not updating for so long, let me clean it up