@shekhei 2017-09-19
Ok, latest changes made