@zbjornson 2017-09-22
@indutny What do you think of landing this awesome PR? Will be nice to have this core part of node.js sped up!