@indutny 2017-09-22
This PR is awesome indeed!

I'll try to find time to review it this week. Sorry for not doing it sooner?

@bnoordhuis @mscdex your help would be very appreciated.