@mscdex 2017-09-22
It would be interesting to see how much of a difference there would be in using NEON on ARM for the parts using intrinsics. I might try that this weekend.