@Kronuz 2017-09-23
Clang doesn’t define `__GNUC__`, yet it does support `emmintrin.h`, please we need to remove the `__GNUC__` checks and just leave `__SSE2__` to make this code efficient also with clang.