@zbjornson 2017-09-23
Clang does define `__GNUC__` actually. (Have you found some version that doesn't?)