@Kronuz 2017-09-23
Well, here it wasn’t compiling that part and it was because of that `__GNUC__` check. I'll dig deeper.