@zbjornson 2017-09-23
Did you have `-march=[something with at least SSE2` or `-msse2`?