@mscdex 2017-09-23
Ok I just added a NEON-based implementation and I see ~13% improvement when testing on an Odroid C2:

Before:

```
Benchmark result:
Took 29.264072 seconds to run
170857.968750 req/sec
```

After:

```
Benchmark result:
Took 25.583895 seconds to run
195435.453125 req/sec
```