@mscdex 2017-09-23
FWIW here is my patch on top of this PR as it currently stands:

```diff
--- http_parser.c	2017-09-23 16:50:15.330221304 -0400
+++ http_parser-neon.c	2017-09-23 16:52:40.372395441 -0400
@@ -443,7 +443,7 @@
     }                                                         \
     ch = *++p;                                                \
     ++parser->nread;                                          \
-  } else                                                      \
+  }
 
 
 /**
@@ -485,19 +485,17 @@
 
 const char* find_crlf(const char* p, const char* data, size_t len);
 
+#ifndef USE_INTRISICS_CRLF
+# define USE_INTRISICS_CRLF 1
+#endif
+
+#if USE_INTRISICS_CRLF && (defined(_MSC_VER) || (defined(__SSE2__) && defined(__GNUC__)))
 #if defined(_MSC_VER) // SSE2 is baseline
 #include <intrin.h>
 #define __builtin_ctz _tzcnt_u32
 #elif defined(__SSE2__) && defined(__GNUC__)
 #include <emmintrin.h>
 #endif
-
-#ifndef USE_INTRISICS_CRLF
-# define USE_INTRISICS_CRLF defined(_MSC_VER) || (defined(__SSE2__) && defined(__GNUC__))
-#endif
-
-#if USE_INTRISICS_CRLF
-
 int32_t get_crlf_mask(const char* p) {
   /* [ c, 0, 0, 0, 0, 0 .. 0 ] */
   const __m128i vCR = _mm_set1_epi8(0x0a);
@@ -539,9 +537,63 @@
   }
   return p;
 }
+#elif USE_INTRISICS_CRLF && defined(__ARM_NEON)
+#include <arm_neon.h>
+const uint8_t __attribute__ ((aligned (16))) _powers_array[16] =
+  { 1, 2, 4, 8, 16, 32, 64, 128, 1, 2, 4, 8, 16, 32, 64, 128 };
+uint8x16_t _powers;
+uint8_t _powers_init = 0;
+uint16_t get_crlf_mask(const char* p) {
+  /* [ c, 0, 0, 0, 0, 0 .. 0 ] */
+  const uint8x16_t vCR = vdupq_n_u8(0x0d);
 
-#else
+  /* [ c, 0, 0, 0, 0, 0 .. 0 ] */
+  const uint8x16_t vLF = vdupq_n_u8(0x0a);
+
+  uint8x16_t v1, v2;
+  v1 = *(uint8x16_t*)p;
+  v2 = vceqq_u8(vCR, v1);
+  v1 = vceqq_u8(vLF, v1);
+  v2 = vorrq_u8(v1, v2);
+
+  // Compute the mask from the input
+  uint64x2_t mask = vpaddlq_u32(vpaddlq_u16(vpaddlq_u8(vandq_u8(v2, _powers))));
+
+  // Get the resulting bytes
+  uint16_t output;
+  vst1q_lane_u8((uint8_t*)&output + 0, (uint8x16_t)mask, 0);
+  vst1q_lane_u8((uint8_t*)&output + 1, (uint8x16_t)mask, 8);
+
+  return output;
+}
+
+const char* find_crlf(const char* p, const char* data, size_t len) {
+  const char* lastp = MIN(data + len, HTTP_MAX_HEADER_SIZE + p);
+
+  int32_t result = 0;
+
+  size_t alignment = (long) p & 15;
+  p -= alignment;
+
+  result = get_crlf_mask(p + 16) << 16 | get_crlf_mask(p);
+  result = (result >> alignment) << alignment;
 
+  if (!result) {
+    while (!result && lastp >= p + 32) {
+      p += 32;
+      result = get_crlf_mask(p + 16) << 16 | get_crlf_mask(p);
+    }
+    if (!result) {
+      return data + len;
+    }
+  }
+  p += __builtin_ctz(result);
+  if (p >= lastp) {
+    return data + len;
+  }
+  return p;
+}
+#else
 const char* find_crlf(const char* p, const char* data, size_t len) {
   const char* p_cr;
   const char* p_lf;
@@ -563,7 +615,6 @@
   }
   return p;
 }
-
 #endif
 
 /* Our URL parser.
@@ -2181,6 +2232,13 @@
   parser->type = t;
   parser->state = (t == HTTP_REQUEST ? s_start_req : (t == HTTP_RESPONSE ? s_start_res : s_start_req_or_res));
   parser->http_errno = HPE_OK;
+#if USE_INTRISICS_CRLF && defined(__ARM_NEON)
+  // Only create the powers list once
+  if (_powers_init == 0) {
+    _powers = vld1q_u8(_powers_array);
+    _powers_init = 1;
+  }
+#endif
 }
 
 void
```