@shekhei 2017-09-24
@Kronuz actually I have tested it on clang too, it works, maybe you can share which version you are using?