@mscdex 2017-09-24
I've updated my patch a bit.

I'm also wondering if the fallback code might benefit from a SWAR implementation over the current dual `memchr()` solution?