@shekhei 2017-09-25
made some changes to ensure not going before or after requested bounds within findcrlf, did a bench, almost no drop in performance