@shekhei 2017-10-18
hmm... I have added some tests a while ago, any comments? or anything else that I should change?