@shekhei 2017-10-19
Ah, it seems like it is stricter with gcc, i have modified, lets see how's the build