@shekhei 2017-10-19
Ok made some changes according to comments