@mscdex 2017-10-19
Still green. LGTM if the suggestion made by @bnoordhuis about `HTTP_MAX_HEADER_SIZE + p` is not a blocker.