@bnoordhuis 2018-01-19
There were some instances of UB last time I looked.  Have those been fixed?

I'm curious if there are actual speedups in larger programs.  Mixing SIMD and non-SIMD code sometimes performs worse than just non-SIMD code.